Extract CompactionJobResources
